feat: implement core controllers and standardize ID types to Integer across layers in backend

+++ medily-backend/src/main/java/com/medily/backend/dto/appointment/AppointmentRequestDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class AppointmentRequestDTO {
     @NotNull
-    private Long doctorId;
+    private Integer doctorId;
     @NotNull
     private String date;// yyyy-MM-dd
     @NotNull
