+++ medily-backend/src/main/java/com/medily/backend/dto/medical/MedicalRecordRequestDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class MedicalRecordRequestDTO {
     @NotNull
-    private Long patientId;
+    private Integer patientId;
     @NotBlank
     private String diagnosis;
     private String notes;
