+++ medily-backend/src/main/java/com/medily/backend/repository/ClinicRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClinicRepository extends JpaRepository<Clinic, Long> {
+public interface ClinicRepository extends JpaRepository<Clinic, Integer> {
     List<Clinic> findByCity(String city);
 }
