+++ medily-backend/src/main/java/com/medily/backend/repository/PrescriptionRequestRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PrescriptionRequestRepository extends JpaRepository<PrescriptionRequest, Long> {
-    List<PrescriptionRequest> findByPatientPatientId(Long patientId);
-    List<PrescriptionRequest> findByPharmacyPharmacyId(Long pharmacyId);
-    List<PrescriptionRequest> findByPatientPatientIdAndStatus(Long patientId, PrescriptionRequest.Status status);
-    List<PrescriptionRequest> findByPharmacyPharmacyIdAndStatus(Long pharmacyId, PrescriptionRequest.Status status);
+public interface PrescriptionRequestRepository extends JpaRepository<PrescriptionRequest, Integer> {
+    List<PrescriptionRequest> findByPatientPatientId(Integer patientId);
+    List<PrescriptionRequest> findByPharmacyPharmacyId(Integer pharmacyId);
+    List<PrescriptionRequest> findByPatientPatientIdAndStatus(Integer patientId, PrescriptionRequest.Status status);
+    List<PrescriptionRequest> findByPharmacyPharmacyIdAndStatus(Integer pharmacyId, PrescriptionRequest.Status status);
 }
