+++ medily-backend/src/main/java/com/medily/backend/service/custom/PostService.java
@@ -6,7 +6,7 @@ import com.medily.backend.dto.post.PostResponseDTO;
 import java.util.List;
 
 public interface PostService {
-    PostResponseDTO createPost(Long doctorUserId, PostCreateRequestDTO request);
+    PostResponseDTO createPost(Integer doctorUserId, PostCreateRequestDTO request);
     List<PostResponseDTO> getAllPosts();
-    void deletePost(Long postId);
+    void deletePost(Integer postId);
 }
