+++ medily-backend/src/main/java/com/medily/backend/service/custom/UserService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface UserService {
     List<UserResponseDTO> getAllUsers();
-    UserResponseDTO getUserById(Long id);
-    void deleteUser(Long id);
+    UserResponseDTO getUserById(Integer id);
+    void deleteUser(Integer id);
 }
